1. router config 2. img path fixes 3. loginInProgress
